Refactor new user flow code with keyboard aware text area, black next arrow
