+++ navigation/BottomTabNavigator.tsx
@@ -21,8 +21,8 @@ import TabTwoScreen from "../screens/TabTwoScreen";
 import { TouchableOpacity } from "react-native";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import { createStackNavigator } from "@react-navigation/stack";
-import selectFamiliarLang from "../screens/selectFamiliarLang";
-import selectInterests from "../screens/selectInterests";
+import selectFamiliarLang from "../screens/new_users/SelectFamiliarLang";
+import selectInterests from "../screens/new_users/selectInterests";
 import selectStartingLanguage from "../screens/selectStartingLang";
 import selectYouthElderly from "../screens/selectYouthElderly";
 import selfIntroduction from "../screens/selfIntroduction";
